Reorganize packages, fix app dark mode handling
