+++ app/build.gradle
@@ -15,7 +15,7 @@ android {
         applicationId "com.rokoblak.chatbackup"
         minSdk 24
         targetSdk 33
-        versionCode 1
+        versionCode 2
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
